Fix file_permissions_etc_audit_rulesd in Image Mode
The rule file_permissions_etc_audit_rulesd fails in a scan executed
after VM deployment of a CentOS Stream 9 bootable container image
hardened with the STIG profile. The rule requires that all files in the
`/etc/audit/rules.d/*.rules` directory need to have mode 0600.  However,
the scan report shows 2 files with mode 0640. This rule passed during
the build of the bootable container image. Therefore, the offending
files were created after the rule is evaluated.  These files are created
by a remediation of a different rule
`audit_rules_kernel_module_loading_delete`.  We can fix the problem by
setting the expected mode at the time of creating these files.  The file
mode set by `bash_fix_audit_syscall_rule` was inconsistent: on line 1768
we set it to 0600, but in this case we just removed permissions of the
others. With this fix the file mode value in the macro will be
consistently set to 0600.