Support subdirectories in other_content
After this change, the keys in other_content will be allowed to be not
only file names but also they can be relative file paths, including
directories. In other words, the other_content will be able to supply
content also in directories and subdirectories. The directory structure
will be preserved in the tarball.