Resolve Jinja in shared test content
The files located in `tests/shared` will not be copied directly
to the tarball to be uploaded. Instead, before copying them,
Jinja variables will be substituted in them. This change will allow us
to use conditions and product properties and other Jinja code in files
in the `tests/shared` directory.