Workaround package_nfs-utils_removed issue on bootable containers
The rule is affected by https://issues.redhat.com/browse/RHEL-74244
which causes that RHEL 10 bootable container hardened with STIG profile
fails to build. Therefore, until this issue is resolved, we will have
a special RHEL 10 only remediation which workarounds the problem on
bootable containers.

Moreover, the `nfs-utils` package is removed as a dependency of
`gssproxy` package so the same workaround is also needed for the
rule `package_gssproxy_removed`.