Expand install_vm with default osinfo for unreleased distros.
Whenever a distro is being develop there might be a need to install
this distro and virt-install requires an osinfo that in this case cannot
be detected because the distro has not been released yet. Usually the
previous major version osinfo data works just fine for the installation
to succeed.