Use new template for firewall package/service rules on Ubuntu 24.04
This change modifies the firewall package/service rules to use the templates
`..._guard_var` introduced in #11818 to selectively install the firewall
that is chosen by the var_network_filtering_service

It also fixes the platform applicability on Ubuntu 24.04 since it
both required firewalld and required that conflicting services
be disabled when installing packages. This interfered with the
logic introduced in the new templates and could result in a
package/service not be installed/enabled.

For example, if the user selected 'nftables' as their firewall
using the new template and variable, the rule package_nftables_installed
would still be marked as not applicable because the ufw service is enabled
by default on some installations. The proposed solution removes the
applicability check and installs the package depending only on the choice of
var_network_filtering_service, irrespective of the status of the ufw service.