Change authselect base profile for custom profiles
In more recent versions of authselect the local profile was introduced
to replace the minimal profile and it is the default authselect profile
for some products. However the local profile does not include all
features required by Benchmarks. This commit ensures the sssd profile
is used as reference to create custom profiles only in cases local
profile is currently selected.

Signed-off-by: Marcus Burghardt <maburgha@redhat.com>