Rule: mount_option_boot_efi_nosuid, remove "uefi" platform
The rule does not care about /sys/firmware/efi, so we should not
use existence of this directory as a guardian.

Moreover following the logic of "mount_option" template the rule
will check and remediate options only if mount point
is active/present in /etc/fstab. That's enough to not interfere
with non-UEFI systems.