replace wrong variable with the correct one
in RHEL 8 STIG, there was a wrong variable used for the rule set_password_hashing_min_rounds_logindefs.
The variable was actually not included in the profile at all and therefore its default value of 5000 was used.
However, stig requires values of 100000.
Also, the variable var_password_pam_unix_rounds was used in the file, but in fact it is not used by any rule in the profile.It got removed to reduce confusion.