Fix ocp4_content to use aarch64 instead of arm64
The ocp4_content Dockerfile determines the system architecture using
`uname -m`. For the ARM64 case, we should be comparing that to `aarch64`
not `arm64`. This commit updates the Dockerfile accordingly.

Without this, the Compliance Operator fails to parse the content from
the content container image because the resulting datastream isn't
copied into the /content/ directory.