Fix accounts_password_pam_retry
The rule `accounts_password_pam_retry` failed after kickstart or
Anaconda installation. The problem is that the rule passes
in the initial scan but fails in the final scan because in the
initial scan the rule `enable_authselect` fails and the remediation
for `enable_authselect` made `accounts_password_pam_retry` fail
but remediation for `accounts_password_pam_retry` isn't executed
because it passed in the initial scan.

Fixes: #12277

Used test: `/hardening/anaconda/anssi_bp28_high` on RHEL 9