Allow multiple bind package names
Starting from RHEL 9.6, the actual RPM package that gets installed
if an user runs the `dnf install bind` command isn't `bind` but
it's `bind9.18`. However, we need to keep the old name as well,
so that the rule works on RHEL 9 older than 9.6. We will use the
new feature of the `package_removed` template that allows us
to specify a list of packages in the `pkgname` parameter.