Improve platform processing in Automatus
The function `common.matches_platform` is inefficient because
it needs to load all CPEs. This function is called for each
test scenario which slows down Automatus runs. It works on
CPE level, but using CPEs isn't needed for processing the
`platform` header of test scenarios. We will remove this
function and replace its usage by already existing function
`utils.is_applicable_for_product` that is already used
to process platforms in remediations and SCEs.