OCPBUGS-52885: Exclude FIO notification check on ARM64
Since FIO isn't supported on ARM, we can disable the notification check
rule on that platform, too.

This commit also updates the profile to use `platform` instead of
`platforms`, since `platforms` is deprecated and overly complex for
what we need (it requires that you use a list of one conditions). If we
use `platform` in the rule, then we should also use `platform` in the
profile filter_rules.