Add renameat2 syscall to audit rules
The audit_file_deletion_events rule checks for the `renameat`
syscall. However, there is a similar syscall `renameat2` which
should be checked as well. We don't have a rule for it so in
this commit we will create a new rule and add `renameat2` syscall
everywhere where `renameat` syscall is used.