OCPBUGS-53041: Fix multi-platform support for OVS rules
The approach we're using with OVAL to select rules based on conditions
doesn't work for multi-arch checks. Instead, the rule always passes
even when it's checking a platform that has `openvswitch` for the group
owner instead of `hugetlbfs`.

This commit attempts to simplify the rule by removing the complexity of
OVAL and relying on individual rules, one for checking group ownership
for openvswitch and the other for hugetlbfs. It also uses the platform
consistently, instead of using platforms, which expects a list of
conditions. Using platform is more straightforward, and incorporates the
architecture CPE into the evaluation at runtime.