OCPBUGS-53041: Separate file_groupowner_ovs_conf_db by arch
The approach we're using with OVAL to select rules based on conditions
doesn't work for multi-arch checks. Instead, the rule always passes
even when it's checking a platform that has `openvswitch` for the group
owner instead of `hugetlbfs`.

This commit attempts to simplify the rule by removing the complexity of
OVAL and relying on individual rules, one for each supported
architecture.