Fix the watch format in audit_rules_mac_modification_etc_selinux
The rule audit_rules_mac_modification_etc_selinux configures
an audit watch for the "/etc/selinux" directory. If modern style
watches are used, we need to pass the "path" parameter in a directory
format because the template needs to generate a different audit watch
for directories. They need to contain "-F dir" instead of "-F path".