OCPBUGS-54486: Ensure platform is set appropriately when filtering rules
The filter_rules construct evaluates python to help organize profiles
based on a set of criteria, like `platform`. However, if you set
platform in the profile filter_rules and also use it in the rule, you'll
get an error like the following:

  TypeError: argument of type 'NoneType' is not iterable

This is because the rule.platform is None, and should at least be an
empty list. This commit checks that condition on a copy of the rule
while doing the evaluation.