OCPBUGS-54486: Use platform instead of platforms
We can model the platform applicability checks using platform instead of
platforms, which has some additional logic that doesn't work as well
with boolean logic. It also only supports a list of one.

The subtle differences between the two is confusing, and we should be
able to do everything with platform.