OCPBUGS-54212: Adjust api-server-tls-cipher-suite to include 4.18 suites
OCP 4.18 comes with some updated ciphers in for the OpenShift API
server. The rule we have for check these ciphers failed because the new
ciphers were not in a blessed list of ciphers.

However, they are recommended and were introduced with RFC 8446:

https://datatracker.ietf.org/doc/html/rfc8446

This commit updates the rule to check for three additional ciphers:

    TLS_AES_128_GCM_SHA256
    TLS_AES_256_GCM_SHA384
    TLS_CHACHA20_POLY1305_SHA256

References:

https://ciphersuite.info/cs/TLS_AES_128_GCM_SHA256/
https://ciphersuite.info/cs/TLS_AES_256_GCM_SHA384/
https://ciphersuite.info/cs/TLS_CHACHA20_POLY1305_SHA256/