Fix not-applicable tests in template pam_account_password_faillock
Tests in this template use the macro `tests_init_faillock_vars`
to define individual test parameters and values based on the
required state (correct, strict, lenient_high, lenient_low).

Two test scenarios, which should be not-applicable, were not properly
implemented and resulted in error:
- lenient_high + no upper boundary
- lenient_low + no lower boundary

This fix inserts an additional `# platform = Not Applicable` to
the top of the test script for those cases.

Fixes issue #13319