Remove rule enable_authselect from RHEL10
Authselect is the default package to configure PAM service files, and
it's enabled by default and at the same time it's impossible to disable.
Therefore, the rule enable_authselect is superfluous and we will remove
it from all RHEL 10 profiles. We will keep the rule as a part of the
hidden default profile.

Resolves: https://issues.redhat.com/browse/OPENSCAP-4691