Update platform
These rules don't exactly depend on `pam` but to be more precise
they depend on `libpwquality`. We will change the platform because
we have moved these rules from the `pam` component file to the
`libpwquality` component file but at the same time the component
sanity test requires that rules with the `package` platform must
be listed in the respective component file.