Update 10-ansible.jinja to exclude autofs
Normally, directories mounted via autofs would show as nfs or nfs4 fstype; however, when an autofs entry exists for a mount point which does not yet exist or is not yet being served from the remote server, this presents as the autofs fstype. Not excluding the autofs fstype causes errors to be thrown when attempting find commands within the directories, although realistically they shouldn't be candidates for find commands to begin with.