Improve bash/ansible_ensure_pam_module_option macros
Make sure that bash/ansible_ensure_pam_module_option can handle situation in which option with value is not set correctly

Current implementation was failing in case the pam option(i.e. deny) is set but to an empty value ,i.e.
instead of `deny=5`, if we have the `deny= ` or even `deny=-3` the regex will fail to match the option