Fix platform in test for grub2_password
In rule `grub2_password`, test scenario `invalid_username.fail.sh`
shouldn't be applicable on OL7, OL8, OL9 and RHEL8 because the test
scenario checks for `superusers` variable set but the OVAL on these
products doesn't contain a check for `superusers`. As a result the rule
unexpectedly passed on these products in this test scenario.