Use JSON for storing resolved items
The build system resolves ("compiles") items and saves them to files in
the build directory in the first phase of the build process and then
loads these files to load items in later phases of the build process.
Currently, YAML format is used as a format of these interim artifacts.
This commit changes the format to JSON. The reason is that Python JSON
library is faster than Python YAML library. This change affects only the
build process, it doesn't change format of the files in the git
repository.