Stop processing templated remediations by Jinja
Remediations that were generated from templates already have all Jinja
code resolved. Therefore, it isn't necessary to process them with Jinja
for the second time. In `collect_remediations.py`, we will process with
Jinja only static remediations and we will not process with Jinja the
generated remediations. This will speed the build up because there are
many remediations generated from templates and Jinja processing takes a
lot of time.

As a part of this change we will rename all `parse_from_file_with_jinja`
methods in `Remediation` and subclasses.