Remove jinja2 from banner related var file
Remove Jinja2 from banner related var file by generate content in advance.

The reason to remove Jinja2 macros is because trestle-bot (and any other external solution processing CaC/content files) fails to process the files when there are Jinja2 macros on them. In the past we already tried to solve this problem by externally processing macros from rule.yml files but it proved to be very painful and we ended up extending the SSG library to introduce new functions that are more friendly for integration.