Change platform of grub2_password
We will mark this rule with `platform: not container` to make
it not applicable during container image build but applicable
on running RHEL Image Mode system. The reason is that the rule
checks files in /boot/grub2/ which don't exist during the container
image build because they are created at deployment time by `bootupd`
On running RHEL Image Mode system it makes sense to configure the
password for GRUB. The remediation for the rule isn't available,
users need to remediate manually by `grub2-setpassword`.