Fix file_owner template to only change invalid ownerships
The previous non-recursive remediation always changed
the ownership, even when the existing owner was valid.

This is an issue if there are multiple valid values for the owner
(e.g. OWNERS="root|syslog").

The fix checks if the current owner is valid (in OWNERS),
and changes the ownership only if invalid.