Fix file_groupowner template to only change invalid ownerships
The previous non-recursive remediation always changed
the group-ownership, even when the existing group was valid.

This is an issue if there are multiple valid values for the group
(e.g. GID_OR_NAME="root|syslog").

The fix checks if the current group is valid (in GID_OR_NAME),
and changes the group-ownership only if invalid.