Workaround OpenSCAP issue for Image Mode
Use full path for some commands in /usr/sbin in Bash
remediations.

This is a workaround for OpenSCAP issue:
https://github.com/OpenSCAP/openscap/issues/2242
A proper fix would be in OpenSCAP, but it's less likely
to update OpenSCAP in downstream than to update the content.

Effectively, this change will fix some rules that fail when
building a hardened bootable container image:
https://github.com/ComplianceAsCode/content/issues/13550
https://github.com/ComplianceAsCode/content/issues/13551
https://github.com/ComplianceAsCode/content/issues/13552

In future, this problem will be smaller, because starting from
Fedora 42, /usr/sbin is a symlink to /usr/bin, see:
https://fedoraproject.org/wiki/Changes/Unify_bin_and_sbin