Run some shell commands in check mode
Tasks using shell or command module normally don't run in check mode.
Setting the check_mode attribute to false on a task using shell or command module
will cause that this task will be executed also in check mode.
This can't be done on all these tasks but only on tasks that don't
perform any change to the system but they just read a value or
read a status of something (for example grep commands).