OPENSCAP-5940: Fix SCAPVal fail
Rule `set_password_hashing_yescrypt_cost_factor_logindefs`
references XCCDF Value `var_password_yescrypt_cost_factor_login_defs`
which holds a number but the OVAL check references to the variable
as a string. This caused error SRC-38-1 in SCAPVal Validation Report.

We will fix this problem by extending the template
`key_value_pair_in_file`. We will add a new optional parameter
`variable_datatype` which allows us to specify a data type other
than string. By default, this will be a string but in the affected
rule we explicitly define it to be an integer.

Fixes: https://github.com/ComplianceAsCode/content/issues/13545