Support banner with single quote
The value of `dod_short` selector in variable login_banner_text
contains single quotes but the bash remediation uses single
quotes for the variable which conflicts with the single quote
inside the value and leads to incomplete bash remediation.
To change single quotes to double quotes for the variable,
we need to use a different construction than using the
bash_instantiate_variables Jinja macro.

The commit also adds a test scenario that covers this situation.

Resolves: https://issues.redhat.com/browse/OPENSCAP-5403