Fix rule auditd_freq
The rule `auditd_freq` is parametrized by `var_auditd_freq` variable but
the value of the variable wasn't honored by checks and remediations and
is hardcoded to 50 instead. The reason was `auditd_lineinfile` template
didn't support parametrizing by variables. In this commit, we will
enhance the `auditd_lineinfile` template to support variables.  Then, we
will use the enhanced template to pass in the variable in rule
`auditd_freq`.

Resolves: https://issues.redhat.com/browse/RHEL-64013