Change TS from fail to pass
The test scenario `space.fail.sh` in rule `networkmanager_dns_mode`
should be a pass test scenario. The test scenario checks the situation
when there are spaces around equal sign in the key value pair in
`/etc/NetworkManager/NetworkManager.conf`. NetworkManager uses the GLib
KeyFile format where spaces around equal signs are allowed and don't
affect the configuration.
https://networkmanager.dev/docs/api/latest/nm-settings-keyfile.html and
https://docs.gtk.org/glib/struct.KeyFile.html So we should change the
file name to .pass.sh. But, this change will most likely trigger a
misalignement with DISA because the test has been introduced as
alignment with DISA by #11745 so we will have to solve it elsewhere.

Fixes: https://github.com/ComplianceAsCode/content/issues/13710