Fix sshd_rekey_limit test scenarios
The test /per-rule/oscap/3/sshd_rekey_limit/correct_value_directory.pass
fails. The reason of the fail is that we don't support distributed
configuration in directory for sshd on RHEL 8.  This is signalized by
variable `sshd_distributed_config` which is a property of each product.
We will make the test scenario applicability based on this variable
which will disable this test scenario on RHEL 8. Also we will do a
similar change on other test scenarios in this rule that test the
configuration in the directory.

Fixes: https://github.com/ComplianceAsCode/content/issues/13720