Fix tests for rule grub2_pti_argument
Some test scenarios for rule `grub2_pti_argument` are failing.  The
reason is that `pti` is a substring of `options`.  This is caused by
recent change that make the equal sign `=` optional.  This change has
been introduced by
https://github.com/ComplianceAsCode/content/pull/13726. After this
change some regular expressions for the `pti` argument started to match
`options`. We will prevent that by using word bounds marks (`\<` and
`\>`) in regular expressions. Also, as a part of this commit we will fix
coding style.

Fixes: https://github.com/ComplianceAsCode/content/issues/13732