Move service tasks to the beginning
The goal of this change is to enable and disable all systemd services at
the beginning of the Ansible Playbook, after all packages are installed
and uninstalled, before `service_facts` are gathered.

To make the processing in the Ansible Playbook builder easier for us we
will slightly change the templates `service_enabled` and
`service_disabled`. We will wrap all tasks in these templates to a block
and we will mark this block by a special tag.  We will us this special
tag to easily identify this block in the Ansible Playbook builder
script.