Use UBI minimal instead of UBI micro
The micro image is a subset of the minimal image, and doesn't have
openssl packages. This means check-payload will fail on images built
from the micro image.

Let's use the minimal image at the expense of a larger image so that
we're including the necessary openssl packages to pass the check-payload
verification for release.