Improve OVAL checks for nss-altfiles
To source group data from /usr/lib/group it isn't sufficient
to configure altfiles in /etc/nsswitch.conf but it's also needed
to install the nss-altfiles package. This means that the OVAL check
in rules file_groupowner_etc_chrony_keys need to be extended
so that both /etc/nsswitch.conf contents and nss-altfiles package
will be checked.