Disambiguate APP-SRG and OCP STIGID
The OCP4 STIGID and APP SRG CTR have the same URI.
This causes the reference that is defined later in the DS to be ignored.

This changes the URI to the parent group "Application Security", which
contains other SRGs, like Application Servers, Database and Web Server.

So there is potential for other collisions.