Update applicability checks for Fedora
On some Fedora systems, the "kernel" RPM package isn't installed because
only the "kernel-core" RPM package is installed. This situation breaks
some of our CPE AL checks and that causes that some rules are evaluated
as notapplicable or some remediations don't execute. In this patch,
we will change the checked package to "kernel-core".