Remove custom Ansible remediation from service_pcscd_enabled
This Ansible remediation isn't idempotent because it contains an
unwanted call of the "ansible.builtin.command" module.  It can be easily
replaced by the remediation generated from the service_enabled template.
The rule had a custom Ansible remeditaion as a fix for issue
https://github.com/ComplianceAsCode/content/issues/8002.  However, this
issue is no longer reproducible with the templated remeditaion, which
means don't need the custom remediation anymore.

Resolves: https://issues.redhat.com/browse/OPENSCAP-6243