Detect non-existent PATH directories in RHEL 9 CIS
This commit better aligns our RHEL 9 CIS profiles with the RHEL 9 CIS
Benchmark 2.0 requirement 5.4.2.5 "Ensure root path integrity".
CIS Benchmark requires that all directories in the root user's PATH
variable are owned by root and are existing directories. Based on this
requirement we add the rules no_dirs_unowned_by_root and
root_path_all_dirs to the profile selections. Moreover, we enhance OVAL
in root_path_all_dirs so that the check fails if any of the paths
specified in PATH don't exist. We add a new test scenario covering this
situation.

Resolves: https://issues.redhat.com/browse/RHEL-102330