Remove Jinja from when statement
This change adresses a problem reported by ansible-lint that
the when statement contains Jinja expressions. This affects
Ansible remediations of rules using the grub2_bootloader_argument
template if the rule is parametrized by a variable. For example,
rule grub2_l1tf_argument is affected by this problem.

Addressing:
```
no-jinja-when: No Jinja2 in when.
```