Fix crypto policy settings in RHEL CIS profiles
Add a new rule `configure_custom_crypto_policy_cis` that configures multiple
custom crypto sub policy modules for RHEL CIS. The new rule is very
similar to `fips_custom_stig_sub_policy`. It configures new modules for
system wide crypto policies that reduces the set of usable ciphers in
sshd, MACs, and others.

The rule is templated by a new template `crypto_sub_policies` that is
also introduced in this commit so that the code can be reused in other
similar rules.

This change aligns the RHEL CIS profiles in CaC with:
- CIS RHEL 8 Benchmark v4.0.0
- CIS RHEL 9 Benchmark v2.0.0
- CIS RHEL 10 Benchmark v1.0.1

All crypto requirements of the CIS profiles are now covered by this single rule.

Resolves: https://issues.redhat.com/browse/RHEL-111896