Align content with RHEL 10 CIS requirement 4.1.5
This commit aligns our profiles with CIS RHEL 10 Benchmark v1.0.1
requirement 4.1.5. Only the rule firewalld_loopback_traffic_trusted
aligns with the CIS prose. There is nothing in the CIS text that would
justify the selection of firewalld_loopback_traffic_restricted, so this
is removed.  Also, adding a note that the rule
firewalld_loopback_traffic_trusted checks for the default RHEL 10 state.

Resolves: https://issues.redhat.com/browse/OPENSCAP-6131