Start using configure_custom_crypto_policy_cis in RHEL 9 and 10
I have checked the latest versions of RHEL 9 and RHEL 10 CIS Benchmarks
and I found they also use the approach to configure cryptography by
creating custom crypto policy sub modules. They are very similar to the
RHEL 8 bu sometimes with different values.
In this commit, we will start using the rule configure_custom_crypto_policy_cis
in CIS for RHEL 9 and RHEL 10 as well, in a similar way to RHEL 8.