Create rule accounts_password_pam_pwhistory_enforce_for_root
Create rule `accounts_password_pam_pwhistory_enforce_for_root` that
checks for presence of the `enforce_for_root` option in
`/etc/security/pwhistory.conf`. This is different than other rules which
check for the same option but in different files. This rule is a part
of RHEL 10 CIS profiles. This aligns our content with RHEL 10 CIS
Benchmark v1.0.1.

Resolves: https://issues.redhat.com/browse/OPENSCAP-6097