extend the bootc applicability platform to cover both online and offline state
it was discovered that there are two distint cases when we can be in the bootc environment.
The first one is while the bootc image is actually hardened during its creation.
This is manifested by presence of /ostree symlink.
The second case is when we are in already booted bootc image.
This was already covered before, but the file /run/ostree-booted is not present when building the image.
Switching only to check for the /ostree symlink does not work as well, because in case of running bootc image, the /ostree is a directory, not a symlink.
I think this covers both cases well and with reasonable accuracy.