Update rule has_nonlocal_mta for latest CIS
The rule has_nonlocal_mta currently checks for services listening only
on port 25, but the policy checks also for ports 465 and 587.  The
content was already updated for other distro to include this extra
criteria, so we simply enable that part of the code to RHEL.

This change aligns our CIS profiles with latest CIS Benchmarks
on RHEL 8, 9 and 10.

Resolves: https://issues.redhat.com/browse/OPENSCAP-6083