Fix Ansible sysctl template
If the value of the sysctl item contains a special character
it needs to be escaped before using it in a regular expression
or the regular expression won't work correctly. This happens
in rule sysctl_kernel_core_pattern where the value is `|/bin/false`
which contains the `|` character.

This fixes multiple failing test scenarios for Ansible remediation
in rule sysctl_kernel_core_pattern.