Create rule group_server_with_gui_removed
This commit introduces new rule `group_server_with_gui_removed`
and adds it to CIS profile according to RHEL 10 CIS Benchmark
v1.0.1, requirement 2.1.19. Although the requirement title
suggests to remove only the GNOME Display Manager (gdm), the
prose in the requirement instructs us to remove the whole
`Server with GUI` dnf group. Also the remediation described
in the CIS document wants us to remove the `Server with GUI`
using the `dnf groupremove` command.

The problem with this rule is that we aren't able to write
an OVAL check for it. The rpm probes can't provide any information
about dnf groups as that is a concept that doesn't exist on
rpm level and is known only for dnf tools. We don't have any
dnf probe or dnf query in OpenSCAP. Therefore this rule will
have only an SCE check.

Resolves: https://issues.redhat.com/browse/OPENSCAP-6081